Default new users to the Agentic UI and remove its feature flag
