+++ apps/studio/src/ipc-types.d.ts
@@ -99,9 +99,8 @@ type IpcApi = {
 	getPathForFile: ( file: File ) => string;
 };
 
-interface FeatureFlags {
-	enableAgenticUi: boolean;
-}
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type -- no flags in flight; see `src/lib/feature-flags.ts`
+interface FeatureFlags {}
 
 interface BetaFeatures {
 	remoteSession: boolean;
